Split post 1417 oversized slices
